log4j and checking country

+++ src/main/java/com/example/service/impl/UserServiceImpl.java
@@ -3,22 +3,32 @@ package com.example.service.impl;
 import com.example.entity.User;
 import com.example.exception.DAOException;
 import com.example.exception.ServiceException;
+import com.example.repository.CountryRepository;
 import com.example.repository.UserRepository;
+import com.example.repository.impl.CountryRepositoryImpl;
 import com.example.repository.impl.UserRepositoryImpl;
 import com.example.service.UserService;
+import com.example.servlet.MainServlet;
 import com.example.validator.UserValidator;
 import com.example.validator.UserValidatorImpl;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.util.List;
 
 public class UserServiceImpl implements UserService {
 
+    private static final Logger logger = LogManager.getLogger(UserServiceImpl.class);
+
     private final UserRepository userRepository;
+
+    private final CountryRepository countryRepository;
 
     private final UserValidator validator;
 
     public UserServiceImpl() {
         this.userRepository = new UserRepositoryImpl();
+        this.countryRepository = new CountryRepositoryImpl();
         this.validator = new UserValidatorImpl();
     }
 
@@ -32,6 +42,7 @@ public class UserServiceImpl implements UserService {
         try {
             return userRepository.findAll(sortBy, sortType, countryId, search, page, pageSize);
         } catch (DAOException e) {
+            logger.error("Error while getting all users: {}", e.getMessage(), e);
             throw new ServiceException(e);
         }
     }
@@ -42,6 +53,9 @@ public class UserServiceImpl implements UserService {
                 user.getFirstname(), user.getSurname(), user.getBirthDate())) {
             throw new ServiceException("Information is not valid!");
         }
+        if (user.getCountry() == null || countryRepository.findById(user.getCountry().getId()).isEmpty()) {
+            throw new ServiceException("Country is null or did not find!");
+        }
 
         userRepository.findByLogin(user.getLogin())
                 .orElseThrow(() -> new ServiceException("Login is already in use!"));
@@ -49,6 +63,7 @@ public class UserServiceImpl implements UserService {
         try {
             userRepository.save(user);
         } catch (DAOException e) {
+            logger.error("Error while adding a user: {}", e.getMessage(), e);
             throw new ServiceException(e);
         }
     }
@@ -56,8 +71,12 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteById(Long userId) {
         try {
+            if (userId == null) {
+                throw new ServiceException("User ID is required.");
+            }
             userRepository.deleteById(userId);
         } catch (DAOException e) {
+            logger.error("Error while deleting a user: {}", e.getMessage(), e);
             throw new ServiceException(e);
         }
     }
@@ -67,6 +86,7 @@ public class UserServiceImpl implements UserService {
         try {
             return userRepository.getTotalResult(sortBy, sortType, countryId, search);
         } catch (DAOException e) {
+            logger.error("Error while getting a total result of users: {}", e.getMessage(), e);
             throw new ServiceException(e);
         }
     }
